feat: Add menu options for creating new sprites and configuring render
The code changes in the `create_scene.py` and `create_sprite.py` files add menu options for creating new sprites and configuring render. The "New" option is disabled by default, and the "Configure Render" option is also disabled. The menu options are added to the File and Render menus respectively. Additionally, an "About" option is added to the Help menu, which displays information about the FlapKine application when clicked.